Major upgrades
PCL is now filtered by intensities and it works. The
threshold is predefined as an hyperparameter.
New make_dataset script that takes a folder with images
and creates a dir structure compatible with roboflow
with train, test and validation splitting.
Some corrections related with camera info parameters.
Launch files adjusted accordingly.